# priority_green_wave_api/priority_green_wave_api/Model/Catadioptrico.cs
using priority_green_wave_api.DTOs;

namespace priority_green_wave_api.Model
{
    public class Catadioptrico
    {
        public int Id { get; set; }
        public int IdLocalizacao { get; set; }
        public string Nome { get; set; }
    }
}