# priority_green_wave_api/priority_green_wave_api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "priority_green_wave_api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Catadioptrico/ReadCatadioptrico": {
      "get": {
        "tags": [
          "Catadioptrico"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Catadioptrico"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Catadioptrico/CreateCatadioptrico": {
      "post": {
        "tags": [
          "Catadioptrico"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Catadioptrico"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Catadioptrico/UpdateCatadioptrico": {
      "put": {
        "tags": [
          "Catadioptrico"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Catadioptrico"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Catadioptrico/DeleteCatadioptrico": {
      "delete": {
        "tags": [
          "Catadioptrico"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Catadioptrico"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/ReadUsuario": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/CreateUsuario": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/LoginUsuario": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/UpdateUsuario": {
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Usuario/DeleteUsuario": {
      "delete": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdUsuario",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Veiculo/ReadVeiculo": {
      "get": {
        "tags": [
          "Veiculo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdVeiculo",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Veiculo/CreateVeiculo": {
      "post": {
        "tags": [
          "Veiculo"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/VeiculoRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Veiculo/UpdateVeiculo": {
      "put": {
        "tags": [
          "Veiculo"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Veiculo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Veiculo/DeleteVeiculo": {
      "delete": {
        "tags": [
          "Veiculo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "IdVeiculo",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Veiculo/EstadoEmergencia": {
      "get": {
        "tags": [
          "Veiculo"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Rfid",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Catadioptrico": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "LoginRequestDTO": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Usuario": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "dataNascimento": {
          "format": "date-time",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "motoristaEmergencia": {
          "type": "boolean"
        }
      }
    },
    "UsuarioRequestDTO": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "dataNascimento": {
          "format": "date-time",
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "motoristaEmergencia": {
          "type": "boolean"
        }
      }
    },
    "Veiculo": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "placa": {
          "type": "string"
        },
        "rfid": {
          "type": "string"
        },
        "fabricante": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "ano": {
          "format": "int32",
          "type": "integer"
        },
        "tipoVeiculo": {
          "type": "string"
        },
        "veiculoEmergencia": {
          "type": "boolean"
        },
        "estadoEmergencia": {
          "type": "boolean"
        }
      }
    },
    "VeiculoRequestDTO": {
      "type": "object",
      "properties": {
        "placa": {
          "type": "string"
        },
        "rfid": {
          "type": "string"
        },
        "fabricante": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "ano": {
          "format": "int32",
          "type": "integer"
        },
        "tipoVeiculo": {
          "type": "string"
        },
        "veiculoEmergencia": {
          "type": "boolean"
        },
        "estadoEmergencia": {
          "type": "boolean"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}